removed unnecesary class, fixed ftp utils

+++ src/main/java/com/jmc/binaria/sender/util/SenderThread.java
@@ -1,19 +1,22 @@
 package com.jmc.binaria.sender.util;
 
 import java.io.File;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
-import java.util.StringTokenizer;
 
 import javax.mail.Message;
 import javax.mail.MessagingException;
-import javax.mail.Transport;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.jmc.binaria.sender.model.EmailCampaign;
+import com.jmc.binaria.sender.model.Sender;
+import com.jmc.binaria.sender.model.SmtpSettings;
+import com.jmc.binaria.sender.service.EmailCampaignService;
+import com.sun.mail.smtp.SMTPTransport;
 
 /**
  *
@@ -40,65 +43,67 @@ public class SenderThread {
 	private String name;
 	private String direccion;
 	private String nombreDestinatario;
+	private Sender sender;
+	private EmailCampaignService emailCampaignService;
+	private SmtpSettings smtpSettings;
+	private EmailCampaign email;
 
-	public SenderThread(Properties props, Integer idThread, String direccion, String nombreDestinatario, String de, String asunto, String contenidoHtml,
-			String nombreAdjunto, String producto, String rutaAdjunto, boolean debug) throws InterruptedException {
+	public SenderThread(Properties props, Integer idThread, Sender sender, EmailCampaignService emailCampaignService,
+			SmtpSettings smtpSettings) {
+		this.props = props;
+		this.idThread = idThread;
+		this.sender = sender;
+		this.emailCampaignService = emailCampaignService;
+		this.de = smtpSettings.getFrom();
+		this.asunto = smtpSettings.getSubject();
+		this.manejaArchivos = new BinariaArchivo();
+		this.nombreAdjunto = smtpSettings.getAttachmenName();
+
+	}
+
+	public SenderThread(Properties props, Integer idThread, String direccion, String nombreDestinatario, String de,
+			String asunto, String contenidoHtml, String nombreAdjunto, String producto, String rutaAdjunto,
+			boolean debug) throws InterruptedException {
 		this.producto = producto;
 		this.idThread = idThread;
-		this.direccion= direccion;
+		this.direccion = direccion;
 		this.nombreDestinatario = nombreDestinatario;
 		this.de = de;
 		this.asunto = asunto;
 		this.contenidoHtmlTemplate = contenidoHtml;
 		this.nombreAdjunto = nombreAdjunto;
 		this.rutaAdjunto = rutaAdjunto;
-
 		this.manejaArchivos = new BinariaArchivo();
-
-		this.direccionesFallidas = new ArrayList<String>();
-		this.errores = new ArrayList<String>();
-		this.props = props;
-		this.debug = debug;
 
 	}
 
-	public boolean send() throws Exception {
-		Transport t = null;
-		StringBuilder cadenaHeader;
-		File adjunto = new File(rutaAdjunto);
-		StringTokenizer sti = null;
-
-		String id;
+	public void send(EmailCampaign email) {
+		debug = sender.isDebug();
+		email.setSenderId(sender.getId());
+		SMTPTransport t = null;
+		File adjunto = new File(email.getAttachmentPath());
 		byte[] pdfArray = null;
-
-		//sti = new StringTokenizer(adjunto.getName(), "|");
-		//id = sti.nextToken();
-		//ti.nextToken();
-		//nombreDestinatario = sti.nextToken();
-		String contenidoHtml = contenidoHtmlTemplate;
+		String contenidoHtml = email.getContentEmail();
 		try {
-
-			
 			MimeMessage message = null;
-
+			String direccion = email.getAddresses();
+			String nombreDestinatario = email.getNames();
 			if (direccion != null || direccion != "null") {
 				nombreDestinatario = "" + nombreDestinatario.replace("¬", "/");
 				direccion = direccion.split(";")[0];
-				//cadenaHeader = new StringBuilder(producto);
-				//cadenaHeader.append("+");
-
 				if (direccion.contains(";")) {
-					//cadenaHeader.append(direccion.split(";")[0].replace("@", "="));
+					// cadenaHeader.append(direccion.split(";")[0].replace("@", "="));
 				}
 				if (direccion.contains(",")) {
-					//cadenaHeader.append(direccion.split(",")[0].replace("@", "="));
+					// cadenaHeader.append(direccion.split(",")[0].replace("@", "="));
 				} else if (!direccion.contains(";") || !direccion.contains(",")) {
-					//cadenaHeader.append(direccion.replace("@", "="));
+					// cadenaHeader.append(direccion.replace("@", "="));
 				}
-	
+
 				contenidoHtml = contenidoHtml.replace("[sunombre]", nombreDestinatario);
-				
+
 				t = JavaMailManager.conectarSMTP(props, debug);
+				
 				message = JavaMailManager.crearMensaje(contenidoHtml, de, nombreDestinatario + " " + asunto);
 				if (direccion.contains(";")) {
 					for (String str : direccion.split(";")) {
@@ -114,24 +119,61 @@ public class SenderThread {
 
 				pdfArray = manejaArchivos.convertirFileADataBinaria(adjunto);
 				JavaMailManager.agregarAdjunto(message, pdfArray, nombreAdjunto, ".pdf");
-				t.sendMessage(message, message.getAllRecipients());
-				if (adjunto.exists())
-					adjunto.delete();
-				return true;
+
+				if (adjunto.exists()) {
+					if (!emailCampaignService.emailWasSent(email)) {
+						t.sendMessage(message, message.getAllRecipients());
+						String response = t.getLastServerResponse();
+						String esmtpId = (response.contains("as ")) ? response.split("as ")[1] : "ESMP-ID-NO-FOUND";
+						email.setEsmtpId(esmtpId.trim());						
+						email.setWasSent(true);
+						email.setResponse("sender-thread-" + this.idThread + " entrego el correo exitosamente");
+						emailCampaignService.updateSending(email);
+						logger.info(email.getResponse());
+						adjunto.delete();
+					} else {
+						email.setResponse(email.getResponse() + "\r " + "sender-thread-" + this.idThread
+								+ ": se intento enviar un correo marcado como enviado ");
+						emailCampaignService.updateSending(email);
+						logger.error("No se enviara el correo a {} porque estaba marcado como enviado ",
+								email.getAddresses());
+					}
+				} else {
+					throw new Exception("No se encontro archivo para adjuntar para " + email.getAddresses());
+				}
 
 			}
 
 		} catch (Exception e) {
-			e.printStackTrace();
+			email.setWasSent(false);
+			email.setError(e.getMessage());
+			logger.error(email.getError());
+			emailCampaignService.updateSending(email);
 			try {
 				t.close();
 			} catch (MessagingException ex) {
 				ex.printStackTrace();
 			}
-			throw e;
 		}
-		return false;
 
+
+	}
+
+	public void run() {
+		try {
+			this.send(null);
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+
+	}
+
+	public EmailCampaign getEmail() {
+		return email;
+	}
+
+	public void setEmail(EmailCampaign email) {
+		this.email = email;
 	}
 
 }
